+++ app/src/main/java/eu/faircode/email/WebViewEx.java
@@ -80,6 +80,7 @@ public class WebViewEx extends WebView implements DownloadListener, View.OnLongC
         settings.setLayoutAlgorithm(WebSettings.LayoutAlgorithm.TEXT_AUTOSIZING);
 
         settings.setAllowFileAccess(false);
+        settings.setAllowContentAccess(true); // default
         settings.setCacheMode(WebSettings.LOAD_CACHE_ELSE_NETWORK);
         settings.setMixedContentMode(WebSettings.MIXED_CONTENT_ALWAYS_ALLOW);
 
@@ -117,10 +118,18 @@ public class WebViewEx extends WebView implements DownloadListener, View.OnLongC
         WebSettings settings = getSettings();
 
         boolean dark = Helper.isDarkTheme(context);
-        boolean canForce = WebViewEx.isFeatureSupported(WebViewFeature.FORCE_DARK);
-        if (canForce)
-            WebSettingsCompat.setForceDark(settings, dark && !force_light ? FORCE_DARK_ON : FORCE_DARK_OFF);
-        setBackgroundColor(canForce && force_light ? Color.WHITE : Color.TRANSPARENT);
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU ||
+                Helper.getTargetSdk(context) < Build.VERSION_CODES.TIRAMISU) {
+            boolean canForce = WebViewEx.isFeatureSupported(WebViewFeature.FORCE_DARK);
+            if (canForce)
+                WebSettingsCompat.setForceDark(settings, dark && !force_light ? FORCE_DARK_ON : FORCE_DARK_OFF);
+            setBackgroundColor(canForce && force_light ? Color.WHITE : Color.TRANSPARENT);
+        } else {
+            // https://developer.android.com/reference/android/webkit/WebSettings#setAlgorithmicDarkeningAllowed(boolean)
+            // https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-color-scheme
+            settings.setAlgorithmicDarkeningAllowed(dark && !force_light);
+            setBackgroundColor(force_light ? Color.WHITE : Color.TRANSPARENT);
+        }
 
         float fontSize = 16f /* Default */ *
                 (browser_zoom ? 1f : message_zoom / 100f);